backend/requirements.txt
flask==3.1.3
flask-cors==6.0.2
google-genai>=1.73.0
firebase-admin>=7.4.0
python-dotenv>=1.2.0
requests>=2.33.0
